Added "SHIFT" to use more buttons using combinations.
While the shiftkey button is pressed, the buttons of the controller can have a different mapping than the standard mapping .
this allows to control many more functions from the controller. I use it for example to : increase / decrease loglevel, have a dedicated button to start record and a dedicated button to stop record, i also control my camera Pan/tilt using a shifted X and Y axis.


Example of shift implementaiton is provided in the Wiiu definitions. 

Merged  button_up_maps and button_down_maps, using naming convention ( "_1" means down ,and "_0" means up). This brings momre readability.

Pass the mem object to the controller, so that the controller can set a variable that can be used in any other part. 

The controller.py file is retrocompatible with previous versions of manage.py.