Added Menus, and shift options allowing more control via the controller.
Clicking the HOME Button will put the controller in menu mode.
Clicking HOME2 button will go to next menu. 
click HOME1 will display the help for the menu. 

a set of menus are provided as part of the controller part  : SYSTEM / CALIBRATE / CAMERA , but the objective is that each part adds its own menu to allow easy configuration / run of the part. 

The latest change in this release is a bug resolution in the axis mappings.