parts/actuator.py: Add Arduino drive-train support
Arduino board can be used to generate the PWM signals for controlling
steering/throttle of the car.

Standard Firmata sketch needs to be loaded on the Arduino to allow it
to be managed from the host-cpu runnig python app.

Native arduino drive train support is particularly useful for combo
boards like LattePanda with builtin cpu/Arduino.