management/base.py: Calibration switch for Arduino
Arduino calibration values vary from 0 - 180. This is different from
PCA9685 for which the calibration values vary from 0 - 1500.

Add --arduino switch to the calibrate command to accomodate this.

More details in actuators.md -
"Note that 90 is the usual midpoint (i.e. 1.5 ms pulse width at 50 Hz),
so it is recommended to start with 90 and adjust +/- 5 until you figure
the desired range for steering / throttle."