Fix pytorch lightning error (#830)
* Predictions are now scaled between [0, 1] for training and validation

Previously the predictions were between [-1, 1] for velocity and
steering, but PyTorch Lightning v1.2.4 introduced changes that
prevented this from working.

Now scaling to be between [0, 1] for training and inference and
then unscaling the predictions for running.

* Replaced accuracy and precision metrics with mean squared error

It didn't make sense to use accuracy and precision when predicted
values could take on a continious range of values. MSE is a better
metric.

PyTorch Lightning also would throw an error when trying to use
accuracy or precision with non-integer tensors.