√937-import-imageaugmentation-causes-error (#939)
ImageAugmentation is used during training. It is needed around line 467 of complete.py where cfg.TRANSFORMATIONS is used to determine if image augmentations will be applied during training. This breaks on a RaspberryPi when trying to drive because imgaug dependency is not installed. To fix this, we should move the import of ImageAugmentation from the top of complete.py to line 467 inside the if statement where it is needed, so no attempt is made to import it unless it is needed.

Co-authored-by: Ezward <Ezward@github.com>