Change AI throttle Multiplier on the joystick
Changing AI throttle multiplier on the joystick while running the model instead of changing it on the config file everytime.

Currently the function has been mapped only to the PS3 joystick dpad left and dpad right buttons as they were  not being used, but can be mapped to other joysticks if there are unused buttons.