1043 Handle crossing paths in path follow (#1058)
Issue https://github.com/autorope/donkeycar/issues/1043

Update the path_follow template so that it can handle paths that have cross-overs.  The current code tests every point in the path to find the closest point.  This means that it is possible that a point on a cross-over going in the opposite direction may get selected.  The approach to fix this is to make the search start at the most recent closest point and limit the search to a configurable number of points.  So for instance, consider a perfectly symmetrical figure 8 path with 1000 waypoints in the path whose origin is at the top of the figure 8; then the intersection as at point 250 and point 750.  If we limit the search to 250 points then only 1/4 of the path will be searched.  If we have driving to the intersection from then origin, then if we are at the the 250th point.  We would then search from point 250 to point 500, and so we never get close the other intersection point at 750.

To implement this we updated these parts:

Update CTE part
- uses a different algorithm to find the closest point.  
- It allows for an offset to start the search and a maximum number of points to search.
- It allows the number of points to look ahead and look behind when returning the track that should be followed.  This allows the algorithm to adjust smoothing and allows a longer look ahead to handle shorter distances between waypoints and/or faster driving speeds.
- The run method now returns the index of the nearest point so we can keep track of it and use it as the offset when searching for the nearest point.

Update OriginOffset part
- run method takes the index of the closest point and returns it, so that it can reset it to the start of the path when the origin is reset.

We then updated the path_follow.py template to use the updated parts and we updated cfg_path_follow.py to include configuration for number of search points, as well as the number of look ahead and look behind points used to calculate the track to follow.